pyproject.toml
[tool.poetry]
name = "file to jsonl"
version = "1.0"
description = "The File Upload Interface is a user-friendly application that preprocesses, extracts, and augments text. It uses Flask and Flask-SocketIO for real-time communication, and HTML, CSS, and JavaScript for the interface. Modules like document_converter.py, text_preprocessor.py, segment_extract_text.py, and augmented_data.py perform the necessary processing. The application can convert files to text, extract keywords and entities, and apply text augmentation techniques. It exports the processed data in JSONL."
authors = ["Thomas Cox"]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.10"
flask = "2.1.1"
summa = "^1.2.0"
Werkzeug = "2.1.2"
python-docx = "0.8.11"
spacy = "3.1.3"
nltk = "3.6.6"
jsonlines = "2.0.0"
en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz"}
pdfminer-six = "^20221105"
python-magic = "^0.4.27"
gevent = "^22.10.2"
pdfplumber = "^0.9.0"
scikit-learn = "^1.2.2"
transformers = "^4.28.1"
sentencepiece = "^0.1.98"
sacremoses = "^0.0.53"
flask-limiter = "^3.3.0"
typing-extensions = "4.5.0"
openai = "^0.27.5"
gevent-websocket = "^0.10.1"
dnspython = "2.1.0"
torch = "^2.0.0"
eventlet = "^0.33.3"
flask-socketio = "^5.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"